# Pathfinding Changes