--- a/src/map.cpp
+++ b/src/map.cpp
@@ -711,12 +711,12 @@ bool Map::getPathMatching(const Creature& creature, const Position& targetPos, s
 
 	AStarNode* found = nullptr;
 	int32_t bestMatch = 0;
-	uint8_t iterations = 0;
+	uint16_t iterations = 0;
 	AStarNode* n = nodes.getBestNode();
 	while (n) {
 		iterations++;
 
-		if (iterations >= Map::nodeReserveSize) {
+		if (iterations >= Map::maxViewportX * Map::maxViewportY) {
 			return false;
 		}