… src/monster.cpp
… src/otpch.h