Performance enhancements to pathfinding algorithm. Remove new from node construction to help with memory problems.

1) Nodes are stored in each data structure using a hash of their x,y coordinates for quick look up.
2) Added a visited set so we do not repeat nodes.
3) Moved open nodes to a priority queue that stores them based on their f value for faster f value look up. (No more harsh sorting)
4) Added a sightline check to optimize clear paths performance.
5) Optimized heuristic calculation which increased performance on all paths.
6) Fixes #4884