@NRH-AA 2025-06-20
Here are some speed tests based on the new algorithm. When I ran the current one I would see up to ~80k nanoseconds to execute. This also DOES NOT USE: NEW to create nodes so memory leaks shouldn't be a huge issue anymore.

[DISTANCE]: 6 [Calculated In]: 14 Iterations and 9600 nanoseconds
[DISTANCE]: 10 [Calculated In]: 37 Iterations and 20800 nanoseconds
[DISTANCE]: 9 [Calculated In]: 39 Iterations and 21000 nanoseconds
[DISTANCE]: 6 [Calculated In]: 21 Iterations and 13900 nanoseconds
[DISTANCE]: 4 [Calculated In]: 8 Iterations and 6200 nanoseconds
[DISTANCE]: 4 [Calculated In]: 5 Iterations and 5200 nanoseconds
[DISTANCE]: 2 [Calculated In]: 2 Iterations and 1500 nanoseconds
[DISTANCE]: 4 [Calculated In]: 7 Iterations and 5900 nanoseconds
[DISTANCE]: 4 [Calculated In]: 4 Iterations and 3300 nanoseconds
[DISTANCE]: 6 [Calculated In]: 18 Iterations and 9900 nanoseconds
[DISTANCE]: 7 [Calculated In]: 19 Iterations and 9700 nanoseconds
[DISTANCE]: 6 [Calculated In]: 12 Iterations and 10200 nanoseconds
[DISTANCE]: 4 [Calculated In]: 8 Iterations and 8400 nanoseconds
[DISTANCE]: 5 [Calculated In]: 8 Iterations and 6900 nanoseconds
[DISTANCE]: 6 [Calculated In]: 18 Iterations and 13800 nanoseconds
[DISTANCE]: 8 [Calculated In]: 28 Iterations and 19500 nanoseconds
[DISTANCE]: 4 [Calculated In]: 6 Iterations and 5400 nanoseconds
[DISTANCE]: 9 [Calculated In]: 34 Iterations and 33000 nanoseconds
[DISTANCE]: 7 [Calculated In]: 25 Iterations and 19900 nanoseconds
[DISTANCE]: 4 [Calculated In]: 3 Iterations and 3800 nanoseconds
[DISTANCE]: 4 [Calculated In]: 8 Iterations and 7600 nanoseconds
[DISTANCE]: 7 [Calculated In]: 16 Iterations and 19000 nanoseconds
[DISTANCE]: 8 [Calculated In]: 20 Iterations and 13100 nanoseconds
[DISTANCE]: 7 [Calculated In]: 26 Iterations and 14200 nanoseconds
[DISTANCE]: 9 [Calculated In]: 32 Iterations and 18500 nanoseconds
[DISTANCE]: 8 [Calculated In]: 25 Iterations and 12900 nanoseconds
[DISTANCE]: 7 [Calculated In]: 22 Iterations and 11700 nanoseconds
[DISTANCE]: 6 [Calculated In]: 21 Iterations and 12400 nanoseconds
[DISTANCE]: 5 [Calculated In]: 13 Iterations and 12000 nanoseconds
[DISTANCE]: 3 [Calculated In]: 2 Iterations and 1600 nanoseconds
[DISTANCE]: 4 [Calculated In]: 5 Iterations and 4600 nanoseconds
[DISTANCE]: 3 [Calculated In]: 2 Iterations and 1700 nanoseconds
[DISTANCE]: 6 [Calculated In]: 6 Iterations and 3700 nanoseconds
[DISTANCE]: 11 [Calculated In]: 35 Iterations and 19000 nanoseconds
[DISTANCE]: 5 [Calculated In]: 12 Iterations and 11100 nanoseconds
[DISTANCE]: 6 [Calculated In]: 15 Iterations and 11400 nanoseconds
[DISTANCE]: 9 [Calculated In]: 26 Iterations and 15300 nanoseconds
[DISTANCE]: 8 [Calculated In]: 38 Iterations and 19000 nanoseconds
[DISTANCE]: 8 [Calculated In]: 26 Iterations and 14400 nanoseconds
[DISTANCE]: 6 [Calculated In]: 16 Iterations and 13000 nanoseconds
[DISTANCE]: 8 [Calculated In]: 17 Iterations and 12900 nanoseconds
[DISTANCE]: 4 [Calculated In]: 8 Iterations and 6500 nanoseconds
[DISTANCE]: 4 [Calculated In]: 6 Iterations and 3500 nanoseconds
[DISTANCE]: 4 [Calculated In]: 5 Iterations and 3600 nanoseconds
[DISTANCE]: 6 [Calculated In]: 18 Iterations and 10900 nanoseconds
[DISTANCE]: 6 [Calculated In]: 9 Iterations and 6700 nanoseconds
[DISTANCE]: 6 [Calculated In]: 13 Iterations and 12900 nanoseconds
[DISTANCE]: 3 [Calculated In]: 5 Iterations and 11800 nanoseconds
[DISTANCE]: 6 [Calculated In]: 15 Iterations and 13000 nanoseconds
[DISTANCE]: 8 [Calculated In]: 19 Iterations and 11200 nanoseconds
[DISTANCE]: 10 [Calculated In]: 61 Iterations and 34500 nanoseconds
[DISTANCE]: 9 [Calculated In]: 39 Iterations and 21300 nanoseconds
[DISTANCE]: 11 [Calculated In]: 34 Iterations and 16900 nanoseconds
[DISTANCE]: 10 [Calculated In]: 38 Iterations and 18600 nanoseconds
[DISTANCE]: 8 [Calculated In]: 11 Iterations and 6300 nanoseconds
[DISTANCE]: 7 [Calculated In]: 15 Iterations and 8500 nanoseconds
[DISTANCE]: 7 [Calculated In]: 19 Iterations and 9200 nanoseconds
[DISTANCE]: 5 [Calculated In]: 6 Iterations and 5700 nanoseconds
[DISTANCE]: 3 [Calculated In]: 2 Iterations and 2200 nanoseconds
[DISTANCE]: 3 [Calculated In]: 5 Iterations and 5300 nanoseconds
[DISTANCE]: 5 [Calculated In]: 13 Iterations and 9900 nanoseconds
[DISTANCE]: 8 [Calculated In]: 27 Iterations and 15100 nanoseconds
[DISTANCE]: 8 [Calculated In]: 33 Iterations and 16000 nanoseconds
[DISTANCE]: 12 [Calculated In]: 51 Iterations and 23000 nanoseconds
[DISTANCE]: 11 [Calculated In]: 34 Iterations and 30600 nanoseconds
[DISTANCE]: 6 [Calculated In]: 14 Iterations and 27200 nanoseconds
[DISTANCE]: 8 [Calculated In]: 24 Iterations and 21300 nanoseconds
[DISTANCE]: 11 [Calculated In]: 34 Iterations and 25300 nanoseconds
[DISTANCE]: 10 [Calculated In]: 43 Iterations and 25700 nanoseconds
[DISTANCE]: 10 [Calculated In]: 30 Iterations and 17800 nanoseconds
[DISTANCE]: 8 [Calculated In]: 17 Iterations and 14200 nanoseconds
[DISTANCE]: 10 [Calculated In]: 16 Iterations and 10800 nanoseconds
[DISTANCE]: 5 [Calculated In]: 7 Iterations and 6600 nanoseconds
[DISTANCE]: 6 [Calculated In]: 9 Iterations and 7200 nanoseconds
[DISTANCE]: 4 [Calculated In]: 4 Iterations and 4200 nanoseconds
[DISTANCE]: 6 [Calculated In]: 11 Iterations and 6500 nanoseconds
[DISTANCE]: 3 [Calculated In]: 2 Iterations and 2100 nanoseconds
[DISTANCE]: 8 [Calculated In]: 17 Iterations and 10800 nanoseconds
[DISTANCE]: 8 [Calculated In]: 21 Iterations and 11400 nanoseconds
[DISTANCE]: 7 [Calculated In]: 12 Iterations and 16500 nanoseconds
[DISTANCE]: 4 [Calculated In]: 5 Iterations and 5600 nanoseconds
[DISTANCE]: 7 [Calculated In]: 22 Iterations and 15400 nanoseconds
[DISTANCE]: 9 [Calculated In]: 21 Iterations and 12300 nanoseconds
[DISTANCE]: 10 [Calculated In]: 52 Iterations and 31700 nanoseconds
[DISTANCE]: 10 [Calculated In]: 41 Iterations and 20100 nanoseconds
[DISTANCE]: 12 [Calculated In]: 38 Iterations and 19100 nanoseconds
[DISTANCE]: 11 [Calculated In]: 44 Iterations and 19500 nanoseconds
[DISTANCE]: 9 [Calculated In]: 22 Iterations and 10500 nanoseconds
[DISTANCE]: 8 [Calculated In]: 19 Iterations and 9600 nanoseconds
[DISTANCE]: 8 [Calculated In]: 11 Iterations and 6800 nanoseconds
[DISTANCE]: 6 [Calculated In]: 8 Iterations and 4200 nanoseconds
[DISTANCE]: 4 [Calculated In]: 5 Iterations and 3900 nanoseconds
[DISTANCE]: 4 [Calculated In]: 6 Iterations and 5200 nanoseconds
[DISTANCE]: 5 [Calculated In]: 9 Iterations and 5400 nanoseconds
[DISTANCE]: 10 [Calculated In]: 17 Iterations and 8200 nanoseconds
[DISTANCE]: 9 [Calculated In]: 23 Iterations and 11700 nanoseconds
[DISTANCE]: 13 [Calculated In]: 35 Iterations and 20800 nanoseconds
[DISTANCE]: 5 [Calculated In]: 7 Iterations and 5400 nanoseconds
[DISTANCE]: 4 [Calculated In]: 4 Iterations and 11400 nanoseconds
[DISTANCE]: 10 [Calculated In]: 25 Iterations and 22700 nanoseconds
[DISTANCE]: 10 [Calculated In]: 13 Iterations and 17600 nanoseconds
[DISTANCE]: 9 [Calculated In]: 18 Iterations and 21000 nanoseconds
[DISTANCE]: 6 [Calculated In]: 19 Iterations and 19100 nanoseconds
[DISTANCE]: 8 [Calculated In]: 43 Iterations and 38700 nanoseconds
[DISTANCE]: 9 [Calculated In]: 28 Iterations and 21300 nanoseconds
[DISTANCE]: 5 [Calculated In]: 10 Iterations and 9600 nanoseconds
[DISTANCE]: 5 [Calculated In]: 13 Iterations and 12500 nanoseconds