@NRH-AA 2025-06-20
After more testing I realized the issue with how large the reserve needed to be was because I was casting the to size_t. Now we can reserve far less memory and it works as intended.