@NRH-AA 2025-06-21
Alright, the final optimization is including a path check to see if the sightline is clear. If it is we can save a lot of performance ignoring nodes that do not need to be checked. This increases the performance of clear paths anywhere from 2-10x.