@gesior 2025-06-28
> from what I gather new shouldn’t be used when working with pathfinders

It should not affect performance much or at all. I can benchmark it tomorrow on Windows, where both versions of code work.
You can try to allocate it on heap, but I tried few solutions and all failed, when I tried to compile with address sanitizer.

> What is the time being measured there?

The same as before: https://github.com/otland/forgottenserver/pull/4924#issuecomment-2946887911

Results:
- **old algorithm**: 7808 ms (100%)
- new algorithm with memory leak: 8289 ms (106%)
- new algorithm without memory leak: 7416 ms (95%)
- **new algorithm from this PR**: 4350 ms (56%)